requirements.txt
aiogram
python-dotenv
gigachat 
openpyxl 